dcraw modified 9.26 (1.476). Changelog: Support the Blackmagic URSA, Nokia N9, and Photron BC2-HD. Support the Canon EOS 750D, 760D, EOS M, and sRAW/mRAW for EOS 5DS R. When there's a tie for largest raw image, -s says which to decode. Fixed null-byte warnings in TIFF output.

+++ dcraw_indi.c
@@ -53,7 +53,7 @@ extern const float d65_white[3];
 #define SQR(x) ((x)*(x))
 #define LIM(x,min,max) MAX(min,MIN(x,max))
 #define ULIM(x,y,z) ((y) < (z) ? LIM(x,y,z) : LIM(x,z,y))
-#define CLIP(x) LIM(x,0,65535)
+#define CLIP(x) LIM((int)(x),0,65535)
 #define SWAP(a,b) { a ^= b; a ^= (b ^= a); }
 
 /*
